pyproject.toml
[tool.poetry]
name = "childhope"
version = "0.1.0"
description = ""
authors = ["\"\""]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
darts = "^0.31.0"
pandas = "^2.2.3"
ydata-profiling = "^4.12.0"
torchinfo = "^1.8.0"
tqdm = "^4.66.6"
tensorboard = "^2.18.0"
torch-tb-profiler = "^0.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
